% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoband_grch37}
\alias{cytoband_grch37}
\title{Cytoband information (GRCh37)}
\format{
A data frame with 4 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{band}{band name}
}
}
\source{
Ensembl, \url{https://www.ensembl.org}
}
\usage{
cytoband_grch37
}
\description{
A dataset containing cytoband information based on GRCh37.
}
\examples{
data(cytoband_grch37)
}
\references{
Harrison, PW, et al. (2024). Ensembl 2024. Nucleic Acids Research, 52, Pages D891–D899. \doi{10.1093/nar/gkad1049}
}
\keyword{datasets}
