% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cytoband_grch38}
\alias{cytoband_grch38}
\title{Cytoband information (GRCh38)}
\format{
A data frame with 4 columns:
\describe{
  \item{chr}{chromosome}
  \item{start}{start position}
  \item{end}{end position}
  \item{band}{band name}
}
}
\source{
Ensembl, \url{https://www.ensembl.org}
}
\usage{
cytoband_grch38
}
\description{
A dataset containing cytoband information based on GRCh38.
}
\examples{
data(cytoband_grch38)
}
\references{
Harrison, PW, et al. (2024). Ensembl 2024. Nucleic Acids Research, 52, Pages D891–D899. \doi{10.1093/nar/gkad1049}
}
\keyword{datasets}
