% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gene_names_grch37}
\alias{gene_names_grch37}
\title{Gene Names (GRCh37)}
\format{
A data frame with 6 columns:
\describe{
  \item{id}{Gene identifier}
  \item{name}{Gene name}
  \item{chr}{Chromosome}
  \item{start}{Start position}
  \item{end}{End position}
  \item{type}{Gene type}
}
}
\source{
Ensembl, \url{https://www.ensembl.org}
}
\usage{
gene_names_grch37
}
\description{
A dataset containing gene names based on GRCh37.
}
\examples{
data(gene_names_grch37)
}
\references{
Harrison, PW, et al. (2024). Ensembl 2024. Nucleic Acids Research, 52, Pages D891–D899. \doi{10.1093/nar/gkad1049}
}
\keyword{datasets}
