% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GWASCat_graph}
\alias{GWASCat_graph}
\title{GWAScatalog graph}
\format{
An object of igraph class.
}
\source{
GWAS Catalog, \url{https://www.ebi.ac.uk/gwas}
}
\usage{
GWASCat_graph
}
\description{
A graph object of GWAScatalog data. The association between variants and phenotypes are indicated.
}
\examples{
data(GWASCat_graph)
}
\references{
Sollis, E, et al. (2022).  The NHGRI-EBI GWAS Catalog: knowledgebase and deposition resource. Nucleic Acids Research, 51, Pages D977–D985. \doi{10.1093/nar/gkac1010}
}
\keyword{datasets}
